tools/requirements.txt
pyyaml>=6.0
rich>=13.7

